Fix issue and split files for active space methods
