+++ forte/api/forte_python_module.cc
@@ -66,113 +66,15 @@
 
 #include "post_process/spin_corr.h"
 
+#include "forte_python_module.h"
+
 namespace py = pybind11;
 using namespace pybind11::literals;
 
 namespace forte {
 
-// see the files in src/api for the implementation of the following methods
-void export_EPICTensors(py::module& m);
-void export_ActiveSpaceIntegrals(py::module& m);
-void export_ForteIntegrals(py::module& m);
-void export_ForteOptions(py::module& m);
-void export_MOSpaceInfo(py::module& m);
-void export_RDMs(py::module& m);
-
-void export_Determinant(py::module& m);
-void export_String(py::module& m);
-void export_Configuration(py::module& m);
-
-void export_SQOperatorString(py::module& m);
-void export_SparseExp(py::module& m);
-void export_SparseFactExp(py::module& m);
-void export_SparseHamiltonian(py::module& m);
-void export_SparseOperator(py::module& m);
-void export_SparseOperatorList(py::module& m);
-void export_SparseOperatorSimTrans(py::module& m);
-void export_SparseState(py::module& m);
-void export_Determinant2(py::module& m);
-void export_GenCIStringLists(py::module& m);
-void export_GenCIVector(py::module& m);
-void export_StateInfo(py::module& m);
-void export_SigmaVector(py::module& m);
-void export_SparseCISolver(py::module& m);
-void export_ForteCubeFile(py::module& m);
-void export_OrbitalTransform(py::module& m);
-void export_Localize(py::module& m);
-void export_SemiCanonical(py::module& m);
-void export_DavidsonLiuSolver(py::module& m);
-
 void set_master_screen_threshold(double value);
 double get_master_screen_threshold();
-
-/// Export the ActiveSpaceMethod class
-void export_ActiveSpaceMethod(py::module& m) {
-    py::class_<ActiveSpaceMethod, std::shared_ptr<ActiveSpaceMethod>>(m, "ActiveSpaceMethod")
-        .def("compute_energy", &ActiveSpaceMethod::compute_energy)
-        .def("set_quiet_mode", &ActiveSpaceMethod::set_quiet_mode)
-        .def("dump_wave_function", &ActiveSpaceMethod::dump_wave_function)
-        .def("read_wave_function", &ActiveSpaceMethod::read_wave_function);
-}
-
-void export_ActiveSpaceSolver(py::module& m) {
-    py::class_<ActiveSpaceSolver, std::shared_ptr<ActiveSpaceSolver>>(m, "ActiveSpaceSolver")
-        .def("compute_energy", &ActiveSpaceSolver::compute_energy)
-        .def("rdms", &ActiveSpaceSolver::rdms)
-        .def("compute_contracted_energy", &ActiveSpaceSolver::compute_contracted_energy,
-             "as_ints"_a, "max_body"_a,
-             "Solve the contracted CI eigenvalue problem using given integrals")
-        .def("compute_average_rdms", &ActiveSpaceSolver::compute_average_rdms,
-             "Compute the weighted average reference")
-        .def("state_energies_map", &ActiveSpaceSolver::state_energies_map,
-             "Return a map of StateInfo to the computed nroots of energies")
-        .def("set_active_space_integrals", &ActiveSpaceSolver::set_active_space_integrals,
-             "Set the active space integrals manually")
-        .def("set_Uactv", &ActiveSpaceSolver::set_Uactv,
-             "Set unitary matrices for changing orbital basis in RDMs when computing dipoles")
-        .def("compute_dipole_moment", &ActiveSpaceSolver::compute_dipole_moment,
-             "Compute transition dipole moment")
-        .def("compute_quadrupole_moment", &ActiveSpaceSolver::compute_quadrupole_moment,
-             "Compute transition quadrupole moment")
-        .def("compute_fosc_same_orbs", &ActiveSpaceSolver::compute_fosc_same_orbs,
-             "Compute the oscillator strength assuming using same orbitals")
-        .def("state_ci_wfn_map", &ActiveSpaceSolver::state_ci_wfn_map,
-             "Return a map from StateInfo to CI wave functions (DeterminantHashVec, eigenvectors)")
-        .def("state_filename_map", &ActiveSpaceSolver::state_filename_map,
-             "Return a map from StateInfo to wave function file names")
-        .def("dump_wave_function", &ActiveSpaceSolver::dump_wave_function,
-             "Dump wave functions to disk")
-        .def("eigenvectors", &ActiveSpaceSolver::eigenvectors, "Return the CI wave functions");
-
-    m.def("compute_average_state_energy", &compute_average_state_energy,
-          "Compute the average energy given the energies and weights of each state");
-}
-
-void export_MCSCF_2STEP(py::module& m) {
-    py::class_<MCSCF_2STEP>(m, "MCSCF_2STEP")
-        .def("compute_energy", &MCSCF_2STEP::compute_energy, "Compute the MCSCF energy");
-}
-
-void export_Symmetry(py::module& m) {
-    py::class_<Symmetry>(m, "Symmetry")
-        .def(py::init<std::string>())
-        .def("point_group_label", &Symmetry::point_group_label,
-             "Returns the label of this point group")
-        .def("irrep_labels", &Symmetry::irrep_labels, "Returns a vector of irrep labels")
-        .def("irrep_label", &Symmetry::irrep_label, "h"_a, "Returns the label of irrep ``h``")
-        .def("irrep_label_to_index", &Symmetry::irrep_label_to_index, "label"_a,
-             "Returns the index of a given irrep ``label``")
-        .def("nirrep", &Symmetry::nirrep, "Returns the number of irreps")
-        .def(
-            "__repr__",
-            [](const Symmetry& sym) { return "Symmetry(" + sym.point_group_label() + ")"; },
-            "Returns a representation of this object")
-        .def(
-            "__str__", [](const Symmetry& sym) { return sym.point_group_label(); },
-            "Returns a string representation of this object")
-        .def_static("irrep_product", &Symmetry::irrep_product, "h"_a, "g"_a,
-                    "Returns the product of irreps ``h`` and ``g``");
-}
 
 // TODO: export more classes using the function above
 PYBIND11_MODULE(_forte, m) {
@@ -300,7 +202,7 @@ PYBIND11_MODULE(_forte, m) {
     export_ActiveSpaceMethod(m);
     export_ActiveSpaceSolver(m);
 
-    export_MCSCF_2STEP(m);
+    export_MCSCF(m);
     export_ForteIntegrals(m);
 
     export_Symmetry(m);
@@ -320,8 +222,6 @@ PYBIND11_MODULE(_forte, m) {
     export_SparseOperatorList(m);
     export_SparseOperatorSimTrans(m);
     export_SparseState(m);
-
-    export_Determinant2(m);
 
     export_GenCIStringLists(m);
     export_GenCIVector(m);
